AHN tiles in init
